+++ components/wasm-opt-cxx-sys/src/shims.h
@@ -79,11 +79,6 @@ namespace wasm_shims {
   std::unique_ptr<ModuleReader> newModuleReader() {
     return std::make_unique<ModuleReader>();
   }
-
-  void ModuleReader_readText(const std::unique_ptr<ModuleReader> &reader, std::string& filename, Module& wasm) {
-    reader->inner.readText(std::move(filename), wasm);
-  }
-
 }
 
 namespace wasm_shims {
@@ -231,8 +226,16 @@ namespace wasm_shims {
       inner.setAll();
     }
 
-    void set(uint32_t feature) {
-      inner.set(feature);
+    void set(uint32_t feature, bool val) {
+      inner.set(feature, val);
+    }
+
+    bool has(const WasmFeatureSet& features) const {
+      return inner.has(features.inner);
+    }
+
+    uint32_t as_int() const {
+      return inner;
     }
   };