+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 ## 0.111.0
 
 - Upgraded to Binaryen 111.
-- Binaryen now enables the `SignExt` and `MutableGlobal` features by default,
+- [Fixed bugs in feature selection via the API](https://github.com/brson/wasm-opt-rs/issues/123).
+- Binaryen now enables the `SignExt` and `MutableGlobals` features by default,
   which are also enabled in the LLVM backend.
   In the future Binaryen will align its default feature selection with the LLVM backend.
   To get the same feature selection as Binaryen 110, call
 
   ```rust
-      opts.disable_feature(Feature::SignExt)
-          .disable_feature(Feature::MutableGlobals)
+      opts.mvp_features_only()
   ```
 - The `TypedFunctionReferences` feature has been removed. The CLI still accepts
   `--enable-typed-function-references` and `--disabled-type-function-references`